Moved authntication and token services to the new _service folder.
