+++ src/app/pages/user/login/login.component.ts
@@ -15,7 +15,7 @@ import { ValidationsServiceService } from 'src/app/services/validation/validatio
 // import { UserService } from 'src/app/services/users.service';
 import { ApiService } from 'src/app/_services/api-service/api.service';
 import { Subscription } from 'rxjs';
-import { AuthService } from 'src/app/services/authorization-services/authentication/auth.service';
+import { AuthService } from 'src/app/_services/authentication-service/auth.service';
 
 @Component({
     selector: 'app-login',
